+++ src/QuickMarcEditor/QuickMarcEditorRows/ContentField/ContentField.js
@@ -1,6 +1,7 @@
 import React, {
   useRef,
   useLayoutEffect,
+  useCallback,
 } from 'react';
 import PropTypes from 'prop-types';
 
@@ -16,6 +17,12 @@ export const ContentField = ({
   ...props
 }) => {
   const ref = useRef();
+
+  const processSubfieldFocus = useCallback(({ target }) => {
+    const end = target.value.length;
+
+    target.setSelectionRange(end, end);
+  }, []);
 
   useLayoutEffect(() => {
     if (ref.current) {
@@ -35,6 +42,7 @@ export const ContentField = ({
       input={input}
       inputRef={ref}
       data-testid={id}
+      onFocus={processSubfieldFocus}
     />
   );
 };